+++ src/senaite/core/upgrade/v02_06_000.py
@@ -31,10 +31,13 @@ from plone.namedfile import NamedBlobFile
 from Products.Archetypes.utils import getRelURL
 from Products.CMFCore.permissions import View
 from senaite.core import logger
+from senaite.core.api.catalog import del_index
 from senaite.core.api.catalog import reindex_index
 from senaite.core.catalog import ANALYSIS_CATALOG
 from senaite.core.catalog import CLIENT_CATALOG
 from senaite.core.catalog import CONTACT_CATALOG
+from senaite.core.catalog import REPORT_CATALOG
+from senaite.core.catalog import SAMPLE_CATALOG
 from senaite.core.catalog import SETUP_CATALOG
 from senaite.core.config import PROJECTNAME as product
 from senaite.core.interfaces import IContentMigrator
@@ -44,6 +47,7 @@ from senaite.core.setuphandlers import setup_other_catalogs
 from senaite.core.upgrade import upgradestep
 from senaite.core.upgrade.utils import UpgradeUtils
 from senaite.core.upgrade.utils import copy_snapshots
+from senaite.core.upgrade.utils import del_metadata
 from senaite.core.upgrade.utils import delete_object
 from senaite.core.upgrade.utils import permanently_allow_type_for
 from senaite.core.upgrade.utils import uncatalog_object
@@ -69,6 +73,8 @@ REMOVE_AT_TYPES = [
     "SamplePreservations",
     "SampleTemplate",
     "SampleTemplates",
+    "Manufacturer",
+    "Manufacturers",
 ]
 
 CONTENT_ACTIONS = [
@@ -1007,6 +1013,53 @@ def migrate_samplepoints_to_dx(tool):
     logger.info("Convert SamplePoints to Dexterity [DONE]")
 
 
+@upgradestep(product, version)
+def migrate_manufacturers_to_dx(tool):
+    """Converts existing manufacturers to Dexterity
+    """
+    logger.info("Convert Manufacturers to Dexterity ...")
+
+    # ensure old AT types are flushed first
+    remove_at_portal_types(tool)
+
+    # run required import steps
+    tool.runImportStepFromProfile(profile, "typeinfo")
+    tool.runImportStepFromProfile(profile, "workflow")
+
+    # get the old container
+    origin = api.get_setup().get("bika_manufacturers")
+    if not origin:
+        # old container is already gone
+        return
+
+    # get the destination container
+    destination = get_setup_folder("manufacturers")
+
+    # un-catalog the old container
+    uncatalog_object(origin)
+
+    # Mapping from schema field name to a tuple of
+    # (accessor, target field name, default value)
+    schema_mapping = {
+        "title": ("Title", "title", ""),
+        "description": ("Description", "description", ""),
+    }
+
+    # migrate the contents from the old AT container to the new one
+    migrate_to_dx("Manufacturer", origin, destination, schema_mapping)
+
+    # copy snapshots for the container
+    copy_snapshots(origin, destination)
+
+    # remove old AT folder
+    if len(origin) == 0:
+        delete_object(origin)
+    else:
+        logger.warn("Cannot remove {}. Is not empty".format(origin))
+
+    logger.info("Convert Manufacturers to Dexterity [DONE]")
+
+
 def migrate_samplepoint_to_dx(src, destination=None):
     """Migrates a Sample Point to DX in destination folder
 
@@ -1210,3 +1263,15 @@ def setup_user_profile(tool):
     update_workflow_role_mappings(LABCONTACT_WORKFLOW, brains)
 
     logger.info("Setup User Profile [DONE]")
+
+
+def remove_creator_fullname(tool):
+    """Remove getCreatorFullName from catalogs
+    """
+    logger.info("Removing getCreatorFullName from catalogs ...")
+
+    del_index(SAMPLE_CATALOG, "getCreatorFullName")
+    del_metadata(SAMPLE_CATALOG, "getCreatorFullName")
+    del_metadata(REPORT_CATALOG, "getCreatorFullName")
+
+    logger.info("Removing getCreatorFullName from catalogs [DONE]")